Change Lesson to abstract class and add concrete subclass TemporaryLesson
